+++ supabase/lib/storage_client.py
@@ -1,23 +1,11 @@
-from typing import Dict
+from deprecation import deprecated
+from storage3 import SyncStorageClient
+from storage3._sync.file_api import SyncBucketProxy
 
-from supabase.lib.storage.storage_bucket_api import StorageBucketAPI
-from supabase.lib.storage.storage_file_api import StorageFileAPI
 
+class SupabaseStorageClient(SyncStorageClient):
+    """Manage storage buckets and files."""
 
-class SupabaseStorageClient(StorageBucketAPI):
-    """
-    Manage the storage bucket and files
-    Examples
-    --------
-    >>> url = storage_file.create_signed_url("something/test2.txt", 80)  # signed url
-    >>> loop.run_until_complete(storage_file.download("something/test2.txt")) # upload or download
-    >>> loop.run_until_complete(storage_file.upload("something/test2.txt","path_file_upload"))
-    >>> list_buckets = storage.list_buckets()
-    >>> list_files = storage_file.list("something")
-    """
-
-    def __init__(self, url: str, headers: Dict[str, str]):
-        super().__init__(url, headers)
-
-    def StorageFileAPI(self, id_: str) -> StorageFileAPI:
-        return StorageFileAPI(self.url, self.headers, id_)
+    @deprecated("0.5.4", "0.6.0", details="Use `.from_()` instead")
+    def StorageFileAPI(self, id_: str) -> SyncBucketProxy:
+        return super().from_(id_)